% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_NRA.R
\name{get_csv_paths}
\alias{get_csv_paths}
\title{Combines files from datraw folder into Rds file.}
\usage{
get_csv_paths(path)
}
\arguments{
\item{string}{Alternative name than the one the function creates.}
}
\value{
A tibble of all csvs in folder, saved as an Rds file.
}
\description{
Combines files from datraw folder into Rds file.
}
\examples{
x <- "my_csv_path"
get_csv_paths(x) \%>\% get_data()

}
