% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_NRA.R
\name{prep_NRFSP}
\alias{prep_NRFSP}
\title{Combines NRFSP data files with key files.}
\usage{
prep_NRFSP(path_dat, path_key, item_pilot = c(80, 81, 82, 83, 84, 85))
}
\arguments{
\item{path_dat}{String pointing to folder containing csv files, or path to single csv file.}

\item{path_key}{String pointing to folder containing csv files, or path to single csv file.}

\item{item_pilot}{List of numbers that correspond pilot item sequence (e.g., 70, 71, 72, 73, 74, 75).}
}
\value{
A comprehensive tibble where every row is a candidate response, but which can be nested to conduct a variety of analysis.
}
\description{
Combines NRFSP data files with key files.
}
