% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FACE.R
\name{prep_FACE}
\alias{prep_FACE}
\title{Combines FACE response files (sdf) with key (csv)}
\usage{
prep_FACE(
  path_dat,
  path_key,
  list_form_id,
  possible_response_options = c("A", "B", "C", "D", "ABCD"),
  passing_score = 60
)
}
\arguments{
\item{string}{list of the names of the forms (e.g., c("Form199", "Form199", "Form198")) that correspond to each file.}
}
\value{
A comprehensive tibble where every row is a candidate response, but which can be nested to conduct a variety of analysis.
}
\description{
Combines FACE response files (sdf) with key (csv)
}
